requirements.txt
fastapi
uvicorn
openai
requests
jinja2
python-dotenv